pyproject.toml
[tool.poetry]
name = "digiprod-gen"
version = "0.1.0"
description = "Digital Product Generator is a streamlit app that allows users to create digital products such a Print on Demand (POD) designs."
authors = ["fteutsch <florian.teutsch@inovex.de>"]
readme = "README.md"
packages = [{include = "digiprod_gen"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.25.0"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
accelerate = "^0.21.0"
bitsandbytes = "^0.40.2"
openai = "^0.27.8"
opencv-python = "^4.8.0.74"
numpy = "^1.25.1"
selenium = "^4.10.0"
seleniumbase = "^4.16.2"
replicate = "^0.9.0"
stability-sdk = "^0.8.4"
click = "^8.1.7"
pydantic-yaml = "^1.1.1"
selenium-wire = "^5.1.0"
llm-prompting-gen = "^0.1.4"
pillow = "^10.1.0"


[tool.poetry.group.backend.dependencies]
fastapi = "0.103.2"
httpx = "^0.25.0"
uvicorn = "^0.23.2"
python-multipart = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start_digiprod_gen = 'digiprod_gen.app:start_digiprod_gen'